# 后端OpenAPI草案_v1.yaml
openapi: 3.0.3
info:
  title: ChuanYi Backend API (Draft V1)
  version: 1.0.0-draft
  description: |
    Draft API contract for ChuanYi app.
    Scope includes DIY flow, order flow, and added goods/cart flow.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/sku/list:
    get:
      tags: [Component]
      summary: Get component SKU list
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: pageNo
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseComponentSkuPage'
...
  /api/admin/component/categories:
    get:
      tags: [Admin]
      summary: List component categories
      parameters:
        - in: query
          name: withCount
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/category/save:
    post:
      tags: [Admin]
      summary: Save component category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                id:
                  type: integer
                  format: int64
                parentId:
                  type: integer
                  format: int64
                categoryType:
                  type: string
                  enum: [BEAD, SPACER, PENDANT, OTHER]
                name:
                  type: string
                sortNo:
                  type: integer
                status:
                  type: string
                  enum: [ENABLED, DISABLED]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/category/{id}/status:
    post:
      tags: [Admin]
      summary: Update component category status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ENABLED, DISABLED]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/category/{id}:
    delete:
      tags: [Admin]
      summary: Delete component category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/sku/list:
    get:
      tags: [Admin]
      summary: List component sku
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: salesStatus
          schema:
            type: string
            enum: [ON_SALE, OFF_SALE]
        - in: query
          name: pageNo
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/sku/create:
    post:
      tags: [Admin]
      summary: Create component sku
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, name, price]
              properties:
                skuCode:
                  type: string
                categoryId:
                  type: integer
                  format: int64
                name:
                  type: string
                specText:
                  type: string
                beadDiameterMm:
                  type: number
                price:
                  type: number
                stockQty:
                  type: integer
                stockWarnQty:
                  type: integer
                imageUrl:
                  type: string
                salesStatus:
                  type: string
                  enum: [ON_SALE, OFF_SALE]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/sku/{skuId}/save:
    post:
      tags: [Admin]
      summary: Save component sku
      parameters:
        - in: path
          name: skuId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skuCode:
                  type: string
                categoryId:
                  type: integer
                  format: int64
                name:
                  type: string
                specText:
                  type: string
                beadDiameterMm:
                  type: number
                price:
                  type: number
                stockQty:
                  type: integer
                stockWarnQty:
                  type: integer
                imageUrl:
                  type: string
                salesStatus:
                  type: string
                  enum: [ON_SALE, OFF_SALE]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/sku/{skuId}/status:
    post:
      tags: [Admin]
      summary: Update component sku status
      parameters:
        - in: path
          name: skuId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [salesStatus]
              properties:
                salesStatus:
                  type: string
                  enum: [ON_SALE, OFF_SALE]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/sku/batch-status:
    post:
      tags: [Admin]
      summary: Batch update component sku status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skuIds, salesStatus]
              properties:
                skuIds:
                  type: array
                  items:
                    type: integer
                    format: int64
                salesStatus:
                  type: string
                  enum: [ON_SALE, OFF_SALE]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/component/sku/{skuId}:
    delete:
      tags: [Admin]
      summary: Delete component sku
      parameters:
        - in: path
          name: skuId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        requestId:
          type: string
          example: 4f6a3d1f8b5545bb
        data:
          nullable: true
...
    ComponentSku:
      type: object
      properties:
        skuId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        name:
          type: string
        specText:
          type: string
        beadDiameterMm:
          type: number
          format: double
        price:
          type: number
          format: double
        stockQty:
          type: integer
        salesStatus:
          type: string
          enum: [ON_SALE, OFF_SALE, OUT_OF_STOCK]
        imageUrl:
          type: string
    PageMeta:
      type: object
      properties:
        pageNo:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
    ApiResponseComponentSkuPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/ComponentSku'
                page:
                  $ref: '#/components/schemas/PageMeta'
...